1_graph_creation/functions/agent1_ingest_lines/requirements.txt
efunctions-framework
flask
google-cloud-storage
google-genai
requests
